requirements.txt
# Core dependencies for Autonomous Research Agent
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Image processing for visualization analysis
Pillow>=9.0.0
opencv-python>=4.6.0

# Web scraping and API requests
requests>=2.28.0
beautifulsoup4>=4.11.0

# PDF processing
PyMuPDF>=1.20.0
pdfminer.six>=20220524

# Optional deep learning frameworks
torch>=1.12.0
transformers>=4.20.0

# Data visualization
plotly>=5.10.0

# Jupyter compatibility
ipython>=8.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0 